% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biolink-package.R
\docType{package}
\name{biolink-package}
\alias{biolink}
\alias{biolink-package}
\title{biolink: Create Hyperlinks to Biological Databases and Resources}
\description{
Generate urls and hyperlinks to commonly used biological databases and resources based on standard identifiers. This is primarily useful when writing dynamic reports that reference things like gene symbols in text or tables, allowing you to, for example, convert gene identifiers to hyperlinks pointing to their entry in the 'NCBI' Gene database. Currently supports 'NCBI' Gene, 'PubMed', Gene Ontology, 'KEGG', CRAN and Bioconductor.
}
\author{
\strong{Maintainer}: Aaron Wolen \email{aaron@wolen.com}

}
\keyword{internal}
