% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biolink.r
\docType{package}
\name{biolink}
\alias{biolink}
\alias{biolink-package}
\title{\pkg{biolink} makes it easy to link to biological resources}
\description{
\code{biolink} provides functions for generating urls and hyperlinks to commonly
used online biological databases. This is primarily useful when writing
\href{https://yihui.name/knitr/}{dynamic reports} that reference biological
identifiers either in text or embedded tables, allowing you to, for example,
convert gene identifiers to hyperlinks to the NCBI Gene database.
}
\details{
biolink can generate valid links for the following document types:
\itemize{
\item LaTex
\item HTML
\item Markdown
}
}
\seealso{
the \href{https://bioconductor.org/packages/annotate}{annotate} package
from Bioconductor
}
