% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-links.r
\name{build_link}
\alias{build_link}
\title{Construct hypertext links}
\usage{
build_link(url, text = NULL, title = NULL, format = "markdown")
}
\arguments{
\item{url}{URL where you want the link to point}

\item{text}{displayed text}

\item{title}{link title, often used in tooltips}

\item{format}{generate links using \code{"html"} or \code{"markdown"} syntax}
}
\description{
Convert a bare url to a valid hyperlink formatted for a Markdown, HTML or
LaTeX document.
}
\examples{
build_link("https://r-project.org", "R", "The R Project")
}
