% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvToExpoID.R
\name{ConvToExpoID}
\alias{ConvToExpoID}
\title{Convert different IDs to the unified Biolink ID}
\usage{
ConvToExpoID(eSet = eSet)
}
\arguments{
\item{eSet}{An R6 class object.}
}
\value{
An R6 class object containing the converted ID information.
}
\description{
Convert the IDs of exposure, chemicals, metabolites, genes, or proteins to the unified Biolink ID,
}
\examples{
eSet <- InitBioLink(1)
eSet <- LoadBioLink(eSet = eSet,UseExample = "example#1", FileDirLink = NULL)
eSet <- ConvToExpoID(eSet = eSet)
}
\author{
Mingliang Fang, Bin Wang (corresponding author), Yuting Wang, Ting Wu
}
