% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi-links.r
\name{entrez}
\alias{entrez}
\alias{entrez_link}
\alias{entrez_url}
\title{NCBI Entrez Gene links}
\usage{
entrez_link(id, text = id, title = NULL, format = "html")

entrez_url(id)
}
\arguments{
\item{id}{Entrez Gene ID (e.g., \code{"4609"})}

\item{text}{displayed text, defaults to the supplied \code{id}}

\item{title}{link title, often used in tooltips}

\item{format}{generate links using \code{"html"} or \code{"markdown"} syntax}
}
\description{
Construct a URL or hypertext link to an entry in NCBI's database for
gene-specific information.
}
\examples{
entrez_link("4609", format = "html")
}
\references{
\href{https://www.ncbi.nlm.nih.gov/gene}{NCBI Gene}
}
