% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exit.R
\name{Exit}
\alias{Exit}
\title{End the module analysis}
\usage{
Exit(eSet = eSet)
}
\arguments{
\item{eSet}{An R6 class object.}
}
\description{
End the module analysis
}
\examples{
eSet <- InitBioLink(1)
eSet <- LoadBioLink(eSet = eSet,UseExample = "example#1", FileDirLink = NULL)
eSet <- ConvToExpoID(eSet = eSet)
eSet <- Biolink(eSet = eSet,Mode = "PPI",ChemCas = "default",ChemInchikey = "default",
DiseaseID = "default",MetabolomeID = "default",MetBiospec = "blood",ProteomeID = "default",GenomeID = "default")
eSet <- VizBioLink(eSet = eSet,Mode = "PPI",Layout = "force-directed",Brightness = "dark",Palette = "default2")
Exit(eSet = eSet)
}
\author{
Mingliang Fang, Bin Wang (corresponding author), Yuting Wang, Ting Wu
}
