% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-links.r
\name{go}
\alias{go}
\alias{go_link}
\alias{go_url}
\title{Gene Ontology Link}
\usage{
go_link(id, text = id, title = NULL, format = "html")

go_url(id)
}
\arguments{
\item{id}{gene ontology ID (e.g., \code{"GO:0005539"})}

\item{text}{displayed text, defaults to the supplied \code{id}}

\item{title}{link title, often used in tooltips}

\item{format}{generate links using \code{"html"} or \code{"markdown"} syntax}
}
\description{
Construct a URL or hypertext link to the Gene Ontology (GO) database.
}
\examples{
go_link("GO:0005539", format = "html")
}
\references{
\href{http://amigo.geneontology.org}{Gene Ontology Consortium}
}
