% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-links.r
\name{go_link}
\alias{go_link}
\alias{go_url}
\title{Gene Ontology Link}
\usage{
go_link(id, text = id, title = NULL, format = "html")

go_url(id)
}
\arguments{
\item{text}{displayed text, defaults to the supplied \code{id}}

\item{title}{link title, often used in tooltips}

\item{format}{generate links using \code{"html"} or \code{"markdown"} syntax}
}
\description{
Construct hypertext link to the Gene Ontology (GO) database.
}
\section{Functions}{
\itemize{
\item \code{go_url}: construct bare URL ponting to GO database
}}

\examples{
go_link("GO:0005539", format = "html")
}
\references{
\href{http://amigo.geneontology.org}{Gene Ontology Consortium}
}
