% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitBioLink.R
\name{InitBioLink}
\alias{InitBioLink}
\title{Initialize BioLink module}
\usage{
InitBioLink(123)
}
\arguments{
\item{seednum}{set a random seed number.}
}
\value{
An R6 class object.
}
\description{
Initialize BioLink module analysis. It can generate an R6 class object
integrating all the analysis information.
}
\details{
BioLink module is designed to find the biological relationships between
exposure factors and health outcome. This module adopts the most frequently-used and
authoritative databases, e.g., T3DB, CTD, ToxCast, StringDB, STITCH, KEGG, and GO.
}
\examples{
eSet <- InitBioLink(123)
}
\author{
Mingliang Fang, Bin Wang (corresponding author), Yuting Wang, Ting Wu
}
