% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-links.r
\name{kegg}
\alias{kegg}
\alias{kegg_link}
\alias{kegg_url}
\title{KEGG Pathway Link}
\usage{
kegg_link(id, text = id, title = NULL, format = "html")

kegg_url(id)
}
\arguments{
\item{id}{KEGG pathway ID (e.g., \code{"hsa04915"})}

\item{text}{displayed text, defaults to the supplied \code{id}}

\item{title}{link title, often used in tooltips}

\item{format}{generate links using \code{"html"} or \code{"markdown"} syntax}
}
\description{
Construct a URL or hypertext link to the KEGG Pathway Database
}
\examples{
kegg_link("hsa04915", format = "html")
}
\references{
\href{http://www.kegg.jp/kegg/pathway.html}{KEGG Pathway Database}
}
