% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadBioLink.R
\name{LoadBioLink}
\alias{LoadBioLink}
\title{Load data file for BioLink module}
\usage{
LoadBioLink(PID, UseExample = "default", FileDirIn=NULL)
}
\arguments{
\item{eSet}{An R6 class object.}

\item{UseExample}{chr. Method of uploading data. If "default",user should upload their own data files,
or use "example#1" provided by this module.}

\item{FileDirLink}{chr. the name of the data file provided by user. Remember move it into the working
directory(e.g. "eg_biolink.xlsx").}
}
\value{
An R6 class object containing the input data.
}
\description{
Load data file for BioLink module.
}
\examples{
eSet = InitBioLink()
          eSet = LoadBioLink(PID = res$PID, UseExample = "example#1")
}
\author{
Mingliang Fang, Bin Wang (corresponding author), Yuting Wang, Ting Wu
}
