% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi-links.r
\name{pubmed}
\alias{pubmed}
\alias{pubmed_link}
\alias{pubmed_url}
\title{PubMed links}
\usage{
pubmed_link(id, text = id, title = NULL, format = "html")

pubmed_url(id)
}
\arguments{
\item{id}{PubMed ID (e.g., \code{"23193287"})}

\item{text}{displayed text, defaults to the supplied \code{id}}

\item{title}{link title, often used in tooltips}

\item{format}{generate links using \code{"html"} or \code{"markdown"} syntax}
}
\description{
Construct a URL or hypertext link to a PubMed article
}
\examples{
pubmed_link("23193287", format = "html")
}
\references{
\href{https://www.ncbi.nlm.nih.gov/pubmed/}{NCBI PubMed}
}
