% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizBioLink.R
\name{VizBioLink}
\alias{VizBioLink}
\title{Visualize the biological link}
\usage{
VizBioLink(eSet = eSet,Mode = "PPI",Layout = "force-directed",Brightness = "dark",Palette = "default2")
}
\arguments{
\item{eSet}{An R6 class object.}

\item{Mode}{chr. Method to build the biological link between exposure and disease.
Only PPI is currently supported.}

\item{Method}{chr. The method to draw the net plot.users can select "ggraph" or "cytoscape".}

\item{Layout}{chr. Visualization layout. Available options include "Force-directed" and "Degree-circle".}

\item{Brightness}{chr. Visualization brightness. Available options include "Light" and "Dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1", "default2"
and "Journal". The "Journal" option provides several journal preference styles including cell, nature,
science, lancet, nejm, and jama.}
}
\value{
An R6 class object containing the plot of the biological link.
}
\description{
Visualize the biological link. It should be noted that the corresponding link
has been built by BioLink function prior to using it.
}
\examples{
eSet <- InitBioLink(1)
eSet <- LoadBioLink(eSet = eSet,UseExample = "example#1", FileDirLink = NULL)
eSet <- ConvToExpoID(eSet = eSet)
eSet <- Biolink(eSet = eSet,Mode = "PPI",ChemCas = "default",ChemInchikey = "default",
DiseaseID = "default",MetabolomeID = "default",MetBiospec = "blood",ProteomeID = "default",GenomeID = "default")
eSet <- VizBioLink(eSet = eSet,Mode = "PPI",Layout = "force-directed",Brightness = "dark",Palette = "default2")
}
\author{
Mingliang Fang, Bin Wang (corresponding author), Yuting Wang, Ting Wu
}
