% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biolink.r
\docType{package}
\name{biolink}
\alias{biolink}
\alias{biolink-package}
\title{\pkg{biolink} makes it easy to link to online resources}
\description{
\code{biolink} provides functions for generating urls and hyperlinks to commonly
used biological databases and resources. This is primarily useful when
writing \href{https://yihui.name/knitr/}{dynamic reports} that reference things
like gene symbols in text or tables, allowing you to, for example, convert
gene identifiers to hyperlinks pointing to their entry in the NCBI Gene
database.
}
\section{Notable features}{

\itemize{
\item generate valid hyperlinks for LaTex, HTML and Markdown documents
\item hyperlink text and title attributes can be specified (where possible)
\item \emph{data tags} can be used to insert certain kinds of information from
select databases
}
}

\seealso{
the \href{https://bioconductor.org/packages/annotate}{annotate} package
from Bioconductor
}
