% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkit.r
\docType{package}
\name{linkit}
\alias{linkit}
\alias{linkit-package}
\title{\pkg{linkit} makes it easy to link to online databases}
\description{
\code{linkit} provides functions for generating urls and hyperlinks to commonly
used online biological databases. This is primarily useful when writing
\href{https://yihui.name/knitr/}{dynamic reports} that reference biological
identifiers either in text or embedded tables, allowing you to, for example,
convert gene identifiers to hyperlinks to the NCBI Gene database.
}
\section{Notable features}{

\itemize{
\item generate valid hyperlinks for LaTex, HTML and Markdown documents
\item hyperlink text and title attributes can be specified (where possible)
\item \emph{magic tags} can be used to insert certains kinds of information from
select databases
}
}

\seealso{
the \href{https://bioconductor.org/packages/annotate}{annotate} package
from Bioconductor
}
