% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-links.r
\name{links}
\alias{links}
\alias{go_link}
\alias{kegg_link}
\alias{pubmed_link}
\alias{entrez_link}
\title{Create hyperlinks to online databases based on standard identifiers}
\usage{
go_link(id, text = id, title = NULL, format = "html")

kegg_link(id, text = id, title = NULL, format = "html")

pubmed_link(id, text = id, title = NULL, format = "html")

entrez_link(id, text = id, title = NULL, format = "html")
}
\arguments{
\item{id}{valid identifier for the relevant online database}

\item{text}{displayed text}

\item{title}{link title, often used in tooltips}

\item{format}{generate links using \code{"html"}, \code{"markdown"} or \code{"latex"} syntax}
}
\description{
Create hyperlinks to online databases based on standard identifiers
}
\section{Functions}{
\itemize{
\item \code{go_link}: to Gene Ontology Consortium

\item \code{kegg_link}: to KEGG Pathway Database

\item \code{pubmed_link}: to PubMed based on PMID (PubMed identifier)

\item \code{entrez_link}: to NCBI's database for gene-specific information based
on Entrez ID
}}

\examples{
go_link("GO:0005539", format = "html")
kegg_link("hsa04915", format = "html")
pubmed_link("23193287", format = "html")
entrez_link("4609", format = "html")

}
\references{
\itemize{
\item \href{http://amigo.geneontology.org}{Gene Ontology Consortium}
\item \href{http://www.kegg.jp/kegg/pathway.html}{KEGG Pathway Database}
\item \href{https://www.ncbi.nlm.nih.gov/pubmed/}{NCBI PubMed}
\item \href{https://www.ncbi.nlm.nih.gov/gene}{NCBI Gene}
}
}
