% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-urls.r
\name{urls}
\alias{urls}
\alias{go_url}
\alias{kegg_url}
\alias{pubmed_url}
\alias{entrez_url}
\alias{stock_url}
\alias{cran_url}
\alias{bioc_url}
\title{Construct urls to online resources}
\usage{
go_url(id)

kegg_url(id)

pubmed_url(id)

entrez_url(id)

stock_url(id)

cran_url(id)

bioc_url(id)
}
\arguments{
\item{id}{valid identifier for the relevant online database}
}
\description{
These resource-specific functions return a bare url (i.e., not a hyperlink)
to the relevant online database/resource based on the provided identifier.
}
\section{Functions}{
\itemize{
\item \code{go_url}: for Gene Ontology Consortium

\item \code{kegg_url}: for KEGG Pathway Database

\item \code{pubmed_url}: for PubMed based on PMID (PubMed identifier)

\item \code{entrez_url}: for NCBI's database for gene-specific information based
on Entrez ID

\item \code{stock_url}: for Yahoo Finance based on ticker symbol

\item \code{cran_url}: for R packages available from CRAN

\item \code{bioc_url}: for R packages available from Bioconductor
}}

\examples{
# gene ontology url
go_url("GO:0005539")

# KEGG pathway url
kegg_url("hsa04915")

# PubMed article url
pubmed_url("23193287")
}
